infoblox-demo/requirements.txt
requests>=2.31.0
pyyaml>=6.0
flask>=3.0.0
ipaddress>=1.0.23
pytest>=7.4.0
pytest-cov>=4.1.0
